+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## master
 * chore: the API root `/` now returns the hostname
   [#645](https://github.com/Kong/kuma/pull/645) 
+* feat: FaultInjection entity, support in API and `kumactl` 
+  [#643](https://github.com/Kong/kuma/pull/643)
 * chore: merge mTLS and CA status into one column
   [#637](https://github.com/Kong/kuma/pull/637)
 * fix: `kumactl apply -v ...` support dots in variables name